+++ test/active_storage_test_case.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+require 'test_helpers/blob_helpers'
+
+class ActiveStorageTestCase < ActiveSupport::TestCase
+  include BlobHelpers
+end
